+++ kestra/tofu/variables.tf
@@ -8,3 +8,9 @@ variable "kestra_password" {
   type        = string
   sensitive   = true
 }
+
+variable "state_passphrase" {
+  description = "State encryption passphrase"
+  type        = string
+  sensitive   = true
+}